backend/requirements.txt
# --- Core agent stack ---
langgraph
langchain
langchain-openai
langchain-community
langchain-core 
langchain-mongodb 
langchain-voyageai 
pymongo 
pypdf



# --- “Web search” tool (Tavily API) ---
langchain-tavily

# --- GitHub tool (official SDK) ---
pygithub

# --- Email tool (async SMTP client; you can also use stdlib smtplib) ---


# --- Service / API server ---
fastapi
uvicorn

# --- Config / env ---
python-dotenv

# --- Utilities ---
requests
